add administrator email apis

+++ jcart-administration-back/src/main/java/jcart/administration/back/cjc/dto/in/OrderHistoryCreateInDTO.java
@@ -11,7 +11,7 @@ public class OrderHistoryCreateInDTO {
     private  String comment;
     private  Boolean customerNotified;
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
